[#12] modify code by convention
- [x] context를 적용하여 전역 상태관리에 용이하도록 함
- [x] 컨벤션에 맞게 코드 정리
- [x] useEffect로 InterviewSidebar 컴포넌트 언마운트 될 때 removeEventListener 실행
- [ ] handleStopCaptureClick에 대한 문제 해결
  - handleSetReplayBlob을 따로 실행했을 경우에는 url blob을 잘 가져옴
  - 디버깅 해야 할 것
    - recordedChunks에 녹화된 데이터가 잘 들어갔을 때 handleSetReplayBlob을 실행할 수 있도록 해야 함
      - handleStopCaptureClick에 handleSetReplayBlob함수를 넣어서 실행하는 것으로 문제를 해결할 수 없음
      - addEventListener로 "stop"일 때 handleSetReplayBlob을 실행하도록 해도 문제를 해결할 수 없음
      - WIP
