+++ frontend/src/components/InterviewContent/stylesheet.scss
@@ -5,5 +5,5 @@
   bottom: 0;
   right: 0;
   background-color: whitesmoke;
-  overflow-x: hidden; 
+  overflow: hidden; 
 }
