+++ frontend/src/components/InterviewSidebar/stylesheet.scss
@@ -1,26 +1,23 @@
 .InterviewSidebarContainer {
   height: calc(100% - 220px);
   width: 300px;
-  position: fixed;
   bottom: 0;
   left: 0;
   background-color: gray;
-  overflow-x: hidden; 
+  overflow: hidden; 
 
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: space-around;
 
-  .Webcam {
+  .webcam {
     display: flex;
     flex-direction: column;
     align-items: center;
     border-style: solid;
   }
-
-  
-  .Button {
+  .button {
     width: 150px;
     height: 50px;
   }
