+++ frontend/src/pages/InterviewPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import InterviewSidebarContainer from '../components/InterviewSidebarContainer';
-import InterviewContentContainer from '../components/InterviewContentContainer';
+import InterviewSidebarContainer from '../components/InterviewSidebar';
+import InterviewContentContainer from '../components/InterviewContent';
 
 export default function InterviewPage() {
   return (
@@ -8,5 +8,5 @@ export default function InterviewPage() {
       <InterviewSidebarContainer />
       <InterviewContentContainer />
     </div>
-  )
-};
+  );
+}
