+++ frontend/src/pages/LoginPage.jsx
@@ -5,5 +5,5 @@ export default function Login() {
     <div>
       Login
     </div>
-  )
-};
+  );
+}
